+++ R/LaTeX.R
@@ -35,7 +35,7 @@ to_latex.huxtable <- function (ht, tabular_only = FALSE, ...){
 
   if (! is.na(height <- height(ht))) {
     if (is.numeric(height)) height <- paste0(height, '\\textheight')
-    res <- paste0('\\resizebox*{!}{', height, '}{\n', res, '\n}')
+    res <- tex_glue('\\resizebox*{!}{<< height >>}{\n<< res >>\n}')
   }
 
   cap <- if (! is.na(cap <- caption(ht))) {
@@ -46,10 +46,9 @@ to_latex.huxtable <- function (ht, tabular_only = FALSE, ...){
             center = 'centering',
             right  = 'raggedleft'
           )
-    cap_setup <- paste0('\\captionsetup{justification=', cap_setup, ',singlelinecheck=off}\n')
-    paste0(cap_setup, '\\caption{', cap, '}\n')
+    tex_glue('\\captionsetup{justification=<< cap_setup >>,singlelinecheck=off}\n\\caption{<< cap >>}\n')
   } else ''
-  lab <- if (! is.na(lab <- label(ht))) paste0('\\label{', lab, '}\n') else ''
+  lab <- if (! is.na(lab <- label(ht))) tex_glue('\\label{<< lab >>}\n') else ''
   if (nzchar(lab) && ! nzchar(cap)) warning('No caption set: LaTeX table labels may not work as expected.')
   res <- if (grepl('top', caption_pos(ht))) paste0(cap, lab, res) else paste0(res, cap, lab)
 
@@ -61,7 +60,7 @@ to_latex.huxtable <- function (ht, tabular_only = FALSE, ...){
   )
   res <- paste0(pos_text[1], res, pos_text[2], '\n')
 
-  res <- paste0('\\begin{table}[', latex_float(ht), ']\n', res, '\\end{table}\n')
+  res <- tex_glue('\\begin{table}[<< latex_float(ht) >>]\n<< res >>\\end{table}\n')
 
   return(res)
 }
@@ -95,12 +94,13 @@ huxtable_latex_dependencies <- list(
 #'
 report_latex_dependencies <- function(quiet = FALSE, as_string = FALSE) {
   report <- sapply(huxtable_latex_dependencies, function(ld) {
-    str <- '\\usepackage'
+    package_str <- '\\usepackage'
     if (! is.null(ld$options)) {
-      str <- paste0(str, '[', paste(ld$options, collapse = ','), ']')
+      options_str <- paste(ld$options, collapse = ',')
+      package_str <- tex_glue('<< package_str >>[<< options_str >>]')
     }
-    str <- paste0(str, '{', ld$name, '}\n')
-    str
+    package_str <- tex_glue('<< package_str >>{<< ld$name >>}\n')
+    package_str
   })
   if (! quiet) {
     cat(paste0(report, collapse = ''))
@@ -112,14 +112,9 @@ report_latex_dependencies <- function(quiet = FALSE, as_string = FALSE) {
 
 
 build_tabular <- function(ht) {
-  colspec <- character(ncol(ht))
-  for (mycol in 1:ncol(ht)) {
-    # col type will be redefined when valign is different:
-    colspec[mycol] <- paste0('p{', compute_width(ht, mycol, mycol), '}')
-  }
+  colspec <- sapply(seq_len(ncol(ht)), function (mycol) tex_glue('p{<< compute_width(ht, mycol, mycol) >>}'))
   colspec <- paste0(colspec, collapse = ' ')
-  colspec <- paste0('{', colspec, '}')
-  res <- paste0(colspec, '\n')
+  res <- tex_glue('{<< colspec >>}\n')
 
   display_cells <- display_cells(ht, all = TRUE)
   all_contents  <- clean_contents(ht, type = 'latex')
@@ -149,8 +144,8 @@ build_tabular <- function(ht) {
         padding <- list(left_padding(ht)[drow, dcol], right_padding(ht)[drow, dcol], top_padding(ht)[drow, dcol],
               bottom_padding(ht)[drow, dcol])
         padding <- lapply(padding, function(x) if (is_a_number(x)) paste0(x, 'pt') else x)
-        tpadding <- if (is.na(padding[3])) '' else paste0('\\rule{0pt}{\\baselineskip+', padding[3], '}')
-        bpadding <- if (is.na(padding[4])) '' else paste0('\\rule[-', padding[4], ']{0pt}{', padding[4], '}')
+        tpadding <- if (is.na(padding[3])) '' else tex_glue('\\rule{0pt}{\\baselineskip+<< padding[3] >>}')
+        bpadding <- if (is.na(padding[4])) '' else tex_glue('\\rule[-<< padding[4] >>]{0pt}{<< padding[4] >>}')
         align_str <- switch(align(ht)[drow, dcol],
           left   = '\\raggedright ',
           right  = '\\raggedleft ',
@@ -163,16 +158,17 @@ build_tabular <- function(ht) {
           # reverse of what you think. 'b' aligns the *bottom* of the text with the baseline
           # this doesn't really work for short text!
           ctb <- switch(valign(ht)[drow, dcol], top = 'b', middle = 'c', bottom = 't')
-          contents   <- paste0('\\parbox[', ctb, ']{', width_spec, hpad_loss[1], hpad_loss[2], '}{', contents, '}')
+          contents   <- tex_glue(
+                '\\parbox[<< ctb >>]{<< width_spec >><< hpad_loss[1] >><< hpad_loss[2] >>}{<< contents >>}')
         }
-        hpadding <- lapply(padding[1:2], function (x) if (! is.na(x)) paste0('\\hspace*{', x, '}') else '')
+        hpadding <- lapply(padding[1:2], function (x) if (! is.na(x)) tex_glue('\\hspace*{<< x >>}') else '')
         contents <- paste0(hpadding[1], contents, hpadding[2])
 
         # to create row height, we add invisible \rule{0pt}. So, these heights are minimums.
         # not sure how this should interact with cell padding...
         if (! is.na(row_height <- row_height(ht)[drow])) {
-          if (is.numeric(row_height)) row_height <- paste0(row_height, '\\textheight')
-          contents <- paste0(contents, '\\rule{0pt}{', row_height, '}')
+          if (is.numeric(row_height)) row_height <- tex_glue('<< row_height >>\\textheight')
+          contents <- tex_glue('<< contents >>\\rule{0pt}{<< row_height >>}')
         }
       }
 
@@ -180,13 +176,13 @@ build_tabular <- function(ht) {
       # but only for left hand cells (which will be multicolumn{colspan} )
       if (! is.na(cell_color <- background_color(ht)[drow, dcol]) && mycol == dcol) {
         cell_color <- format_color(cell_color)
-        cell_color <- paste0('\\cellcolor[RGB]{', cell_color, '}')
-        contents <- paste0(cell_color, ' ', contents)
+        cell_color <- tex_glue('\\cellcolor[RGB]{<< cell_color >>}')
+        contents <- tex_glue('<< cell_color >> << contents >>')
       }
 
       if (bottom_left_multirow) {
         # * is 'standard width', could be more specific:
-        contents <- paste0('\\multirow{-', rs, '}{*}{', contents, '}')
+        contents <- tex_glue('\\multirow{-<< rs >>}{*}{<< contents >>}')
       }
 
       if (mycol == dcol) { # first column of cell
@@ -194,24 +190,24 @@ build_tabular <- function(ht) {
         colspec <- if (wrap(ht)[drow, dcol]) {
           pmb <- switch(valign(ht)[drow, dcol], top   = 'p', bottom  = 'b', middle = 'm')
           width_spec <- compute_width(ht, mycol, dcell$end_col)
-          paste0(pmb, '{', width_spec, '}')
+          tex_glue('<< pmb >>{<< width_spec >>}')
         } else {
           switch(align(ht)[drow, dcol], left = 'l', center = 'c', right = 'r')
         }
         # only add left borders if we haven't already added a right border!
-        lb <- if (! added_right_border) v_border(ht, drow, dcol, 'left') else ''
-        rb <- v_border(ht, drow, dcol, 'right')
+        lb <- if (! added_right_border) v_border(ht, myrow, mycol) else ''
+        rb <- v_border(ht, myrow, dcell$end_col + 1) # we need to put the end column here
         added_right_border <- rb != ''
-        contents <- paste0('\\multicolumn{', cs, '}{', lb, colspec, rb, '}{', contents, '}')
+        contents <- tex_glue('\\multicolumn{<< cs >>}{<< lb >><< colspec >><< rb >>}{<< contents >>}')
       }
 
       row_contents[mycol] <- contents
 
     } # next cell
     row_contents <- row_contents[nzchar(row_contents)] # if we've printed nothing, don't print an & for it
-    row_contents <- paste(row_contents, collapse = ' & ')
-    # the 0.5pt avoids nasty pale lines through colored multirow cells
-    res <- paste0(res, row_contents, ' \\tabularnewline[-0.5pt]\n')
+    row_contents <- paste(row_contents, collapse = ' & \n')
+    # should reduce nasty pale lines through colored multirow cells:
+    res <- tex_glue('<< res >><< row_contents >> \\tabularnewline[-0.5pt]\n')
 
     # add top/bottom borders
     res <- paste0(res, build_clines_for_row(ht, myrow))
@@ -221,11 +217,11 @@ build_tabular <- function(ht) {
   width_spec <- if (tenv %in% c('tabularx', 'tabular*', 'tabulary')) {
     tw <- width(ht)
     if (is_a_number(tw)) tw <- paste0(tw, default_table_width_unit)
-    paste0('{', tw, '}')
+    tex_glue('{<< tw >>}')
   } else {
     ''
   }
-  res <- paste0('\\begin{', tenv, '}', width_spec, res, '\\end{', tenv, '}\n')
+  res <- tex_glue('\\begin{<< tenv >>}<< width_spec >><< res >>\\end{<< tenv >>}\n')
 
   return(res)
 }
@@ -256,7 +252,8 @@ compute_width <- function (ht, start_col, end_col) {
 
   if (end_col > start_col) {
     # need to add some extra tabcolseps, two per column
-    cw <- paste0(cw, '+', (end_col - start_col) * 2, '\\tabcolsep')
+    extra_seps <- (end_col - start_col) * 2
+    cw <- tex_glue('<< cw >>+<< extra_seps >>\\tabcolsep')
   }
 
   cw
@@ -264,121 +261,90 @@ compute_width <- function (ht, start_col, end_col) {
 
 build_cell_contents <- function(ht, row, col, contents) {
   if (! is.na(font_size <- font_size(ht)[row, col])) {
-    font_size_pt <- paste0(font_size, 'pt')
-    line_space <- paste0(round(font_size * 1.2, 2), 'pt')
-    contents <- paste0('{\\fontsize{', font_size_pt, '}{', line_space, '}\\selectfont ', contents, '}')
+    line_space <- round(font_size * 1.2, 2)
+    contents <- tex_glue('{\\fontsize{<< font_size >>pt}{<< line_space >>pt}\\selectfont << contents >>}')
   }
   if (! is.na(text_color <- text_color(ht)[row, col])) {
     text_color <- format_color(text_color)
-    contents <- paste0('\\textcolor[RGB]{', text_color, '}{', contents, '}')
+    contents <- tex_glue('\\textcolor[RGB]{<< text_color >>}{<< contents >>}')
   }
-  if (bold(ht)[row, col]) {
-    contents <- paste0('\\textbf{', contents, '}')
-  }
-  if (italic(ht)[row, col]) {
-    contents <- paste0('\\textit{', contents, '}')
-  }
+  if (bold(ht)[row, col])   contents <- tex_glue('\\textbf{<< contents >>}')
+  if (italic(ht)[row, col]) contents <- tex_glue('\\textit{<< contents >>}')
   if (! is.na(font <- font(ht)[row, col])) {
-    contents <- paste0('{\\fontfamily{', font, '}\\selectfont ', contents, '}')
+    contents <- tex_glue('{\\fontfamily{<< font >>}\\selectfont << contents >>}')
   }
-  if ( (rt <- rotation(ht)[row, col]) != 0) {
-    contents <- paste0('\\rotatebox{', rt, '}{', contents, '}')
-  }
+  if ( (rt <- rotation(ht)[row, col]) != 0) contents <- tex_glue('\\rotatebox{<< rt >>}{<< contents >>}')
 
   return(contents)
 }
 
+# row can be from "0" for the top; up to nrow
 build_clines_for_row <- function(ht, row) {
-  # row can be from "0" for the top; up to nrow
-  # add top/bottom borders
   # where a cell is shadowed, we don't want to add a top border (it'll go thru the middle)
   # bottom borders of a shadowed cell are fine, but come from the display cell.
   display_cells <- display_cells(ht, all = TRUE)
-  dcells_this_row <- unique(display_cells[display_cells$row == row, ])
-  this_bottom <- rep('', ncol(ht))
+  display_cells <- display_cells[display_cells$row == row, ]
 
-  blank_line_color <- rep(format_color('white'), ncol(ht)) # white by default, I guess...
-  for (i in seq_len(nrow(dcells_this_row))) {
-    drow <- dcells_this_row[i, 'display_row']
-    dcol <- dcells_this_row[i, 'display_col']
-    end_row <- dcells_this_row[i, 'end_row']
-    end_col  <- dcells_this_row[i, 'end_col']
-    # Print bottom border if we are at bottom of the display cell
-    if (row == end_row) {
-      this_bottom[dcol:end_col] <- h_border(ht, drow, dcol, 'bottom')
-    }
+  blank_line_color <- rep(NA, ncol(ht)) # white by default, I guess...
+  for (i in seq_len(nrow(display_cells))) {
+    dc <- display_cells[i, ]
     # Use color if we are in middle of display cell
-    if (row < end_row & ! is.na(color <- background_color(ht)[drow, dcol])) {
-      blank_line_color[dcol:end_col] <- format_color(color)
+    if (row < dc$end_row) {
+      col <- background_color(ht)[dc$display_row, dc$display_col]
+      blank_line_color[dc$display_col:dc$end_col] <- format_color(col, default = 'white')
     }
   }
-  blanks <- paste0('>{\\arrayrulecolor[RGB]{', blank_line_color, '}}-')
 
-  dcells_next_row <- unique(display_cells[display_cells$row == row + 1, ])
-  next_top <- rep('', ncol(ht))
-  for (i in seq_len(nrow(dcells_next_row))) {
-    drow <- dcells_next_row[i, 'display_row']
-    dcol <- dcells_next_row[i, 'display_col']
-    end_col <- dcells_next_row[i, 'end_col']
-    # are we at the top of this dcell? If not...
-    if (row + 1 != drow) next
-    next_top[dcol:end_col] <- h_border(ht, drow, dcol, 'top')
-  }
-  borders <- this_bottom
-  borders[borders == ''] <- next_top[borders == '']
-
-  if (any(borders != '')) {
-    hhlinechars <- ifelse(borders != '', borders, blanks)
-    # add in |.
-    # For each hhlinechar (1 real column), look at above (and/or below?) left/right borders
-    # | placed at left/right of corresponding - for left/right border;
-    vertlines <- compute_vertical_borders(ht, row)
-    hhlinechars <- paste0(hhlinechars, vertlines[-1], collapse = '')
-    hhlinechars <- paste0(vertlines[1], hhlinechars)
-    hhline <- paste0('\\hhline{', hhlinechars, '}\n')
-    hhline <- paste0(hhline, '\\arrayrulecolor{black}\n') # don't let arrayrulecolor spill over
-    return(hhline)
-  } else {
+  widths <- collapsed_borders(ht)$horiz[row + 1,]
+  if (all(widths == 0)) {
     return('')
+  } else {
+    width <- max(widths)
+    if (! all(widths[widths > 0] == width)) warning("Multiple widths in a single border, using max")
+    colors <- collapsed_border_colors(ht)$horiz[row + 1, ]
+    colors <- sapply(colors, format_color, default = 'black')
+    hhlinechars <- sapply(seq_along(widths), function (x) {
+      col <- if (widths[x] > 0) colors[x] else blank_line_color[x]
+      tex_glue('>{\\arrayrulecolor[RGB]{<< col >>}\\global\\arrayrulewidth=<< width >>pt}-')
+    })
+    vertlines <- compute_vertical_borders(ht, row)
+
+    hhline <- paste0(hhlinechars, vertlines[-1], collapse = '')
+    hhline <- paste0(vertlines[1], hhline)
+    hhline <- tex_glue('\n\n\\hhline{<< hhline >>}\n\\arrayrulecolor{black}\n')
+    return(hhline)
   }
 }
 
+
+# these are inserted into the hhline. They have to have the same arrayrulewidth as the
+# left/right borders of the cell above and below. That is, arrayrulewidth = max(right of above_left,
+# left of above_right, right of below_left, left of below_right).
+# row can be 0 for the top line.
+# returns an ncol(ht) + 1 string array
 compute_vertical_borders <- function (ht, row) {
-  display_cells <- display_cells(ht, all = TRUE)
-  dcells_this_row <- unique(display_cells[display_cells$row == row, c('display_row', 'display_col')])
+  # if we are at the top line, then we'll assume we want the same vertical borders as on the first line below.
+  if (row == 0) row <- 1
+  b_widths <- collapsed_borders(ht)$vert[row, ]
+  b_cols   <- collapsed_border_colors(ht)$vert[row, ]
+  borders <- sapply(seq_along(b_widths), function (x){
+    if (b_widths[x] == 0 ) return('')
+    my_col <- format_color(b_cols[x], default = 'black')
+    tex_glue('>{\\arrayrulecolor[RGB]{<< my_col >>}\\global\\arrayrulewidth=<< b_widths[x] >>pt}|')
+  })
 
-  lbs <- rep('', ncol(ht) + 1)
-  rbs <- rep('', ncol(ht) + 1)
-  for (i in seq_len(nrow(dcells_this_row))) {
-    drow <- dcells_this_row[i, 'display_row']
-    dcol <- dcells_this_row[i, 'display_col']
-    right_col <- dcells_this_row[i, 'end_col'] - 1 # first rbs is always 0
-    lbs[dcol] <- v_border(ht, drow, dcol, 'left')
-    rbs[right_col] <- v_border(ht, drow, dcol, 'right')
-  }
-  lbs[nzchar(lbs)] <- '|' # HACK
-  rbs[nzchar(rbs)] <- '|'
+  return(borders)
+}
 
-  borders <- rbs # these take priority
-  borders[borders == ''] <- lbs
-  borders
+# uses "real" border numbers in "ncol + 1 space"
+v_border <- function (ht, row, col) {
+  width <- collapsed_borders(ht)$vert[row, col]
+  color <- collapsed_border_colors(ht)$vert[row, col]
+  color <- format_color(color, default = 'black')
+
+  tex_glue('!{\\color[RGB]{<< color >>}\\vrule width << width >>pt}')
 }
 
 
-v_border <- function (ht, drow, dcol, side) {
-  width <- get_all_borders(ht, drow, dcol)[side]
-  color <- get_all_border_colors(ht, drow, dcol)[side]
-  if (! width > 0 ) return('')
-  if (is.na(color)) return('|')
-  color <- format_color(color)
-  paste0('!{\\color[RGB]{', color, '}\\vrule}')
-}
-
-h_border <- function (ht, drow, dcol, side) {
-  width <- get_all_borders(ht, drow, dcol)[side]
-  color <- get_all_border_colors(ht, drow, dcol)[side]
-  if (! width > 0 ) return('')
-  if (is.na(color)) color <- 'black' # the default
-  color <- format_color(color)
-  paste0('>{\\arrayrulecolor[RGB]{', color, '}}-')
-}
+# this has to have the ... argument as it's (???) evaluated in the parent frame
+tex_glue <- function (...) glue::glue(..., .open = '<<', .close = '>>', .envir = parent.frame())