+++ tests/testthat/test-huxreg.R
@@ -27,7 +27,8 @@ test_that('huxreg confidence intervals work', {
   glm1 <- glm(I(y > 0) ~ a, dfr, family = binomial)
   library(nnet)
   mn <- nnet::multinom(I(y > 0) ~ a, dfr, trace = FALSE)
-  expect_silent(hr <- huxreg(lm1, lm2, glm1, mn, error_style = "ci", statistics = c('r.squared')))
+  expect_silent(huxreg(lm1, lm2, glm1, mn, error_format = "{conf.low}-{conf.high}", statistics = c('r.squared'),
+        ci_level = 0.95))
 })
 
 test_that('huxreg works with single coefficient', {
@@ -60,13 +61,13 @@ test_that('huxreg number_format works correctly', {
   lm1 <- lm(y ~ a, dfr)
   lm2 <- lm(y ~ b, dfr)
   hr <- huxreg(lm1, lm2, number_format = 4)
-  expect_equal(number_format(hr)[4,2], list(4))
-  expect_equal(number_format(hr)[9,2], list(4))
-  expect_match(to_screen(hr), paste0('\\D', round(coef(lm1)[2], 4) ,'\\D'))
+  expect_equal(number_format(hr)[4, 2], list(4))
+  expect_equal(number_format(hr)[9, 2], list(4))
+  expect_match(to_screen(hr), paste0('\\D', round(coef(lm1)[2], 4), '\\D'))
   hr2 <- huxreg(lm1, lm2, number_format = '%5.3f')
-  expect_equal(number_format(hr2)[4,2], list('%5.3f'))
-  expect_equal(number_format(hr2)[9,2], list('%5.3f'))
-  expect_match(to_screen(hr2), paste0('\\D', sprintf('%5.3f', coef(lm1)[2]),'\\D'))
+  expect_equal(number_format(hr2)[4, 2], list('%5.3f'))
+  expect_equal(number_format(hr2)[9, 2], list('%5.3f'))
+  expect_match(to_screen(hr2), paste0('\\D', sprintf('%5.3f', coef(lm1)[2]), '\\D'))
 })
 
 
@@ -76,8 +77,8 @@ test_that('huxreg borders argument works', {
   dfr$y <- dfr$y + dfr$a
   lm1 <- lm(y ~ a, dfr)
   lm2 <- lm(y ~ b, dfr)
-  hr <- huxreg(lm1, lm2)
-  expect_equivalent(bottom_border(hr)[, 2], c(1, rep(0, 5), 1, rep(0, nrow(hr) - 9), 1, 0))
+  hr <- huxreg(lm1, lm2, borders = .7)
+  expect_equivalent(bottom_border(hr)[, 2], c(.7, rep(0, 5), .7, rep(0, nrow(hr) - 9), .7, 0))
   hr2 <- huxreg(lm1, lm2, borders = FALSE)
   expect_equivalent(bottom_border(hr2)[, 2], rep(0, nrow(hr2)))
 })
@@ -96,4 +97,9 @@ test_that('huxreg stars printed correctly', {
   expect_match(huxreg(lm1, stars = c('@' = 0.1, 'wrong' = 0.05))[[4, 2]], paste0(number_regex, '@\\s*'))
   expect_match(huxreg(lm1, stars = c('wrong' = 0.05, '@' = 0.1))[[4, 2]], paste0(number_regex, '@\\s*'))
   expect_match(huxreg(lm2)[[4, 2]], paste0(number_regex, '\\*\\*\\*\\s*'))
+})
+
+test_that('huxreg works for models without tidy p values', {
+  expect_warning(huxreg(lme4::lmer(Sepal.Width ~ Sepal.Length + (1 | Species), data = iris), statistics = 'nobs'),
+        'p values')
 })